Added imported JSx components from choicefarms
Not breaking at the moment. Need to combine this with the main branch and then create a new Astro instance based on this new deploy.
The reason to create a new Astro instance is because it seems Astro cli is not initialized on the current website. If Astro Cli doesn't work on this new combine, then I might have to add the dependancies such as react manually instead of spending hours trying to figure out why it's not working.